% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_achv_score.R
\name{calculate_achv_score}
\alias{calculate_achv_score}
\title{Calculate the Achievement Score}
\usage{
calculate_achv_score(som_report, award_weights = c(1, 2, 2, 3, 3))
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}

\item{award_weights}{Weights for the five major awards, from Wright Brothers to Spaatz.}
}
\value{
A data.table containing the unit charter number and the achievement score.
}
\description{
Calculate the achievement score for all units in a Squadron of Merit Listing.
}
\details{
By default, the achievement score awards 1 point for every Wright Brothers Award, 2 points for
every Mitchell and Earhart Award, and 3 points for every Eaker and Spaatz Award.

You can pass an alternative vector of weights in to \code{award_weights} if you want to change these
weightings.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
