% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_enc_metric.R
\name{calculate_enc_metric}
\alias{calculate_enc_metric}
\title{Calculate the Encampment Metric}
\usage{
calculate_enc_metric(som_report)
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}
}
\value{
A data.table containing the unit charter number and the encampment metric.
}
\description{
Calculate the encampment metric for all units in a Squadron of Merit Listing.
}
\details{
This metric awards 1 point for every first-time Encampment attendee.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
