% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_enc_score.R
\name{calculate_enc_score}
\alias{calculate_enc_score}
\title{Calculate the Encampment Score}
\usage{
calculate_enc_score(som_report)
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}
}
\value{
A data.table containing the unit charter number and the encampment score.
}
\description{
Calculate the encampment score for all units in a Squadron of Merit Listing.
}
\details{
This score awards 1 point for every first-time Encampment attendee.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
