% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flight_metric.R
\name{calculate_flight_metric}
\alias{calculate_flight_metric}
\title{Calculate the Flight Metric}
\usage{
calculate_flight_metric(som_report)
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}
}
\value{
A data.table containing the unit charter number and the flight metric.
}
\description{
Calculate the flight metric for all units in a Squadron of Merit Listing.
}
\details{
This metric is calculated by multiplying the Flight99 measure by 1, awarding 1 point for a value
of 100\% on the Flight99 measure. The point value is doubled if the orientation flights flag
is set to Yes.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
