% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_growth_metric.R
\name{calculate_growth_metric}
\alias{calculate_growth_metric}
\title{Calculate the Growth Metric}
\usage{
calculate_growth_metric(som_report, focus_area = "balanced")
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}

\item{focus_area}{The focus area for scoring, which defaults to "balanced".}
}
\value{
A data.table containing the unit charter number and the growth metric.
}
\description{
Calculate the growth metric for all units in a Squadron of Merit Listing.
}
\details{
The default growth metric is calculated by awarding 2 points for every cadet who joins or renews,
giving equal weighting to both recruiting and retention. The \code{focus_area} variable can be used to
change this.

Use \code{recruiting} to award 3 points for every new cadet and 1 point for every renewal.
Use \code{retention} to award 1 point for every new cadet and 3 points for every renewal.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
