% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overall_score.R
\name{calculate_overall_score}
\alias{calculate_overall_score}
\title{Calculate the Overall SOM Score}
\usage{
calculate_overall_score(
  som_report_path,
  metric_weights = c(strength = 1, growth = 1, achv = 1, enc = 1, flight = 1),
  growth_focus_area = "balanced",
  achv_award_weights = c(1, 2, 2, 3, 3)
)
}
\arguments{
\item{som_report_path}{A path to the Squadron of Merit Listing report in CSV format.}

\item{metric_weights}{Weights for the five metrics.}

\item{growth_focus_area}{The focus area for growth scoring, which defaults to "balanced".}

\item{achv_award_weights}{Weights for the major cadet awards, from Wright Brothers to Spaatz.}
}
\value{
A data.table containing unit charter number, the metric scores, and the overall score.
}
\description{
Calculate the overall score for all units in a Squadron of Merit Listing.
}
\details{
Each of the metric scores are calculated and a total score is obtained by weighting these scores
equally. If you want to change the weights, pass a named vector to \code{metric_weights}.

You can pass a different growth focus area through using \code{growth_focus_area}. Similarly, you can
pass a different set of award weights to the achievement meetric using \code{achv_award_weights}.
}
