% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_strength_score.R
\name{calculate_strength_score}
\alias{calculate_strength_score}
\title{Calculate the Strength Score}
\usage{
calculate_strength_score(som_report)
}
\arguments{
\item{som_report}{A data.table containing the Squadron of Merit Listing data.}
}
\value{
A data.table containing the unit charter number and the strength score.
}
\description{
Calculate the strength score for all units in a Squadron of Merit Listing.
}
\details{
The strength score is calculated by dividing the starting strength by 12, which assigns 1 point
for the minimum strength as described in CAPR 39-3.

The final result is then standardized using the distribution of the scores to have a mean of 1
and standard deviation of 1.
}
