% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_som_csv.R
\name{load_som_report}
\alias{load_som_report}
\title{Load the Squadron of Merit Listing from CSV}
\usage{
load_som_report(report_path)
}
\arguments{
\item{report_path}{The path to the CSV report file. Local paths are acceptable.}
}
\value{
A data.table containing the Squadron of Merit Listing, with excess fields stripped.
}
\description{
Loads the CSV of the Squadron of Merit Listing report from eServices into R.
}
